Abstracted Tile and Entity classes to inherit off of an abstract Renderable class.
