+++ Makefile
@@ -8,7 +8,7 @@
 EXEC = golf_engine
 
 # Class names
-CLASSES = main Vector2 Entity Tests
+CLASSES = main Tests Vector2 Renderable Tile Entity
 
 # Compiler command
 CC = g++
