+++ src/Renderable.cpp
@@ -1,0 +1,10 @@
+/**
+ * @file Renderable.cpp
+ * @brief This file contains definitions for the Renderable abstract class.
+ *
+ * @author Willow Ciesialka
+ * @date 2023-06-06
+ */
+
+#include "Renderable.hpp"
+using GolfEngine::Renderable;
